Add basic book removal

+++ src/data-access/book-access.ts
@@ -6,10 +6,12 @@ import { book } from "@/db/schema/books";
 import { eq } from "drizzle-orm";
 
 export async function getBooksByUserId(userId: string) {
+  /* Fetches all books for a specific user by their user ID. */
   return await db.select().from(book).where(eq(book.userId, userId));
 }
 
 export async function addBook(bookData: { id: string, title: string; author: string; userId: string }) {
+  /* Add book for a specific user. */
   return await db.insert(book).values({
     ...bookData,
     createdAt: new Date(),
@@ -20,6 +22,7 @@ export async function updateBook(bookId: string, bookData: { title?: string; aut
   return await db.update(book).set(bookData).where(eq(book.id, bookId));
 }
 
-export async function deleteBook(bookId: string) {
-  return await db.delete(book).where(eq(book.id, bookId));
+export async function deleteBook(userId: string, bookId: string) {
+  /* Delete book for a specific user by book ID. */
+  return await db.delete(book).where(eq(book.userId, userId) && eq(book.id, bookId));
 }
